% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.stlplus}
\alias{plot.stlplus}
\title{Lattice plot of the raw, seasonal, trend, and remainder components}
\usage{
\method{plot}{stlplus}(x, scales = list(y = list(relation = "sliced")),
  type = "l", as.table = TRUE, strip = FALSE, strip.left = TRUE,
  between = list(y = 0.5), layout = NULL, ...)
}
\arguments{
\item{x}{object of class \code{"stlplus"}.}

\item{scales, type, as.table, strip, strip.left, between, layout, \ldots}{parameters to be passed to xyplot.}
}
\value{
object of class \code{"trellis"}.
}
\description{
Lattice plot of the raw, seasonal, trend, and remainder components.  If post-trend smoothing was done, these components will be plotted instead of the trend component.
}
\references{
R. B. Cleveland, W. S. Cleveland, J. E.  McRae, and I. Terpenning (1990) STL:  A  Seasonal-Trend  Decomposition  Procedure Based on Loess. \emph{Journal of Official Statistics}, \bold{6}, 3--73.
}
\seealso{
\code{\link{stlplus}}
}

