% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rembycycle}
\alias{plot_rembycycle}
\title{Plot of Remainder Component by Cycle-Subseries}
\usage{
plot_rembycycle(x, col = "darkgray", locol = "black", lolwd = 2,
  xlab = "Time", ylab = "Remainder", ...)
}
\arguments{
\item{x}{object of class \code{"stlplus"}.}

\item{col, locol, lolwd, xlab, ylab, \ldots}{parameters to be passed to \code{xyplot()}. \code{locol} and \code{lolwd} are the line color and width for the loess line.}
}
\value{
object of class \code{"trellis"}.
}
\description{
Plots the remainder component by cycle-subseries with a loess line.
}
\references{
R. B. Cleveland, W. S. Cleveland, J. E.  McRae, and I. Terpenning (1990) STL:  A  Seasonal-Trend  Decomposition  Procedure Based on Loess. \emph{Journal of Official Statistics}, \bold{6}, 3--73.
}
\seealso{
\code{\link{stlplus}}
}

