\name{respsim}
\docType{data}
\alias{respsim}
\title{Simulated daily emergency department respiratory chief complaint counts}
\description{
	A time series of daily counts of patients who checked in to an emergency department with a chief complaint categorized as respiratory.
}
\usage{respsim}
\format{A vector containing 1399 observations.}
\source{Simulated from a model based on real data.}
\references{
   R. P. Hafen, D. E. Anderson, W. S. Cleveland, R. Maciejewski, D. S. Ebert, A. Abusalah, M. Yakout, M. Ouzzani, and S. J. Grannis. Syndromic surveillance: STL for modeling, visualizing, and monitoring disease counts. \emph{BMC Medical Informatics and Decision Making}, 9(1):21, 2009.
}
\keyword{datasets}